refactor(test): split swap-detector tests under SwiftLint body budget
Splits IntraAccountSwapDetectorTests into two suites — predicate-logic
cases stay in IntraAccountSwapDetectorTests; field, order, and self-
send preservation move to SwapDetectorPreservationTests. Trims the
original suite under SwiftLint's 250-line type_body_length budget
without dropping coverage. Also drops the stable-UUID accountId
literal (which required a force unwrap; tests don't depend on its
specific value) in favour of a freshly generated UUID.
